Add support for DOUBLE data type in SchemaBuilder
This commit implements the handling of DOUBLE data type in the SchemaBuilder and related classes. It includes updating the enum in `Schema` to include DOUBLE, modifying `CrateSQL` to map DOUBLE to the corresponding SQL type, and extending `Evolution` to manage schema evolution for DOUBLE fields. Also, new test cases are added to verify these changes.